requirements.txt
Django>=3.2.3,<3.3
django-allauth>=0.43.0,<0.44.0
gunicorn>=20.1.0,<20.2
google-cloud-secret-manager>=2.4.0,<2.5
google-cloud-ndb>=1.8,<1.9
